Query for stop routes separately.
